+++ src/adapters/MailerChainAdapter.php
@@ -3,6 +3,7 @@
 namespace bertoost\mailerchain\adapters;
 
 use bertoost\mailerchain\elements\ChainAdapter;
+use Craft;
 use craft\mail\transportadapters\BaseTransportAdapter;
 use craft\mail\transportadapters\Sendmail;
 use Symfony\Component\Mailer\Transport\AbstractTransport;
@@ -12,6 +13,16 @@ class MailerChainAdapter extends BaseTransportAdapter
     public static function displayName(): string
     {
         return 'Mailer Chain';
+    }
+
+    public function getSettingsHtml(): ?string
+    {
+        return Craft::$app->getView()->renderTemplate(
+            'mailerchain/adapter/settings', [
+                'adapter' => $this,
+                'valid' => ChainAdapter::find()->testSuccess()->exists(),
+            ]
+        );
     }
 
     public function defineTransport(): array|AbstractTransport